Added stricter tests for mode switcher
