+++ test/ModeSwitcher.test.js
@@ -1,23 +1,7 @@
-import ModeSwitcher, {
-  Icon
-} from '../src/saka/Main/Components/ModeSwitcher/index.jsx';
+import ModeSwitcher from '../src/saka/Main/Components/ModeSwitcher/index.jsx';
 import { render } from 'preact-render-spy';
 import { fadedColorMap } from 'lib/colors.js';
 import { h } from 'preact';
-
-describe('Icon component ', () => {
-  it('should render while enabled', () => {
-    const props = {
-      icon: '',
-      color: 'rgba(1,1,1,0.44)'
-    };
-
-    const iconRender = render(<Icon {...props} />);
-    const icon = iconRender.find('#icon');
-
-    expect(icon).toBeTruthy();
-  });
-});
 
 describe('ModeSwitcher component ', () => {
   it('should render tabs with selected tab colored, rest of tabs gray', () => {
@@ -33,15 +17,39 @@ describe('ModeSwitcher component ', () => {
       .at(0)
       .children();
 
+    const recentClosed = modeSwitcher
+      .children()
+      .at(1)
+      .children();
+
     const bookmark = modeSwitcher
       .children()
       .at(2)
       .children();
 
+    const history = modeSwitcher
+      .children()
+      .at(3)
+      .children();
+
+    const recentlyViewed = modeSwitcher
+      .children()
+      .at(4)
+      .children();
+
     expect(tab.attr('icon')).toBe('tab');
     expect(tab.attr('color')).toBe(fadedColorMap.tab);
 
+    expect(recentClosed.attr('icon')).toBe('restore_page');
+    expect(recentClosed.attr('color')).toBe(fadedColorMap.unknown);
+
     expect(bookmark.attr('icon')).toBe('bookmark_border');
     expect(bookmark.attr('color')).toBe(fadedColorMap.unknown);
+
+    expect(history.attr('icon')).toBe('history');
+    expect(history.attr('color')).toBe(fadedColorMap.unknown);
+
+    expect(recentlyViewed.attr('icon')).toBe('timelapse');
+    expect(recentlyViewed.attr('color')).toBe(fadedColorMap.unknown);
   });
 });
